pyproject.toml
[project]
name = "sarm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "equinox>=0.13.2",
    "jax==0.6.0",
    "torch>=2.9.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
convert = [
    "open-clip-torch>=3.2.0",
]
test = [
    "open-clip-torch>=3.2.0",
    "pillow>=12.0.0",
    "pytest>=8.4.2",
]
